Add files via upload
definitely room for improvement here with the chance to combine some of these 'move' methods and cut down on code re use. my solution involved creating a grid, a list of char arrays. looping through these to find the player position represented in the data by '^' and then in my other move methods i'm basically checking the number of dots between the player and the next hash symbol