Add files via upload
day 7 challenge - iterating through a dataset. This was added into a list of long arrays. checked all possible permutations of the values to see if they could be, working from left to right, added or multipled sequentially to get the value of index[0] in each array. then if this was possible the value from index[0] was added to a running tally.